fix user inflation code to handle arbitrary usernames
We were calling $_[0]->isa without first checking to see if it was an
object - this can either be an object or a username. However, the tests
only used usernames that were [a-z] - e.g. 'jim' - which meant that the
->isa was returning false rather than throwing an exception. I've modified
the test to use 3jim and updated the code to correctly check.